Fraction->double in SdrObject::Resize
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/201733
Tested-by: Jenkins

+++ svx/source/svdraw/svdfmtf.cxx
@@ -451,7 +451,7 @@ void ImpSdrGDIMetaFileImport::InsertObj(SdrObject* pObj1, bool bScale)
     {
         if(mbSize)
         {
-            pObj->NbcResize(Point(), maScaleX, maScaleY);
+            pObj->NbcResize(Point(), double(maScaleX), double(maScaleY));
         }
 
         if(mbMov)
