+++ svx/source/svdraw/svdpdf.cxx
@@ -684,7 +684,7 @@ void ImpSdrPdfImport::InsertObj(SdrObject* pObj1, bool bScale)
     {
         if (mbSize)
         {
-            pObj->NbcResize(Point(), maScaleX, maScaleY);
+            pObj->NbcResize(Point(), double(maScaleX), double(maScaleY));
         }
 
         if (mbMov)
